requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cloudinary==1.28.0
dj-database-url==0.5.0
Django==3.2.9
django-heroku==0.3.1
django-phonenumber-field==6.0.0
django-uuidfield==0.5.0
djangorestframework==3.12.4
gunicorn==20.1.0
phonenumberslite==8.12.37
psycopg2==2.9.2
python-decouple==3.5
python-dotenv==0.19.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
whitenoise==5.3.0
